+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Current
 - fixed issue in After hook if browser is not started
+- added I save css property step
+- added I expect css property step